Extract ProductInfoPricing component
