+++ src/ui/states/state-present-offer.svelte
@@ -1,22 +1,12 @@
 <script lang="ts">
-  import Localized from "../localization/localized.svelte";
   import {
-    type NonSubscriptionOption,
     type Product,
-    ProductType,
     type PurchaseOption,
-    type SubscriptionOption,
+    ProductType,
   } from "../../entities/offerings";
-  import { getContext } from "svelte";
-  import { translatorContextKey } from "../localization/constants";
-  import { Translator } from "../localization/translator";
-
-  import { LocalizationKeys } from "../localization/supportedLanguages";
   import type { BrandingAppearance } from "../../entities/branding";
-  import { getTranslatedPeriodLength } from "../../helpers/price-labels";
-  import { getNextRenewalDate } from "../../helpers/duration-helper";
-  import { type Writable } from "svelte/store";
   import ProductInfoHeader from "../components/product-info-header.svelte";
+  import ProductInfoPricing from "../components/product-info-pricing.svelte";
 
   export let productDetails: Product;
   export let purchaseOption: PurchaseOption;
@@ -25,147 +15,20 @@
 
   const isSubscription =
     productDetails.productType === ProductType.Subscription;
-  const subscriptionOption: SubscriptionOption | null | undefined =
-    purchaseOption as SubscriptionOption;
-  const nonSubscriptionOption: NonSubscriptionOption | null | undefined =
-    purchaseOption as NonSubscriptionOption;
-
-  const subscriptionTrial = subscriptionOption?.trial;
-
-  const subscriptionBasePrice = subscriptionOption?.base?.price;
-  const nonSubscriptionBasePrice = nonSubscriptionOption?.basePrice;
-
-  const translator: Writable<Translator> = getContext(translatorContextKey);
-
-  const formattedSubscriptionBasePrice =
-    subscriptionBasePrice &&
-    $translator.formatPrice(
-      subscriptionBasePrice.amountMicros,
-      subscriptionBasePrice.currency,
-    );
-
-  const formattedZeroPrice =
-    subscriptionBasePrice &&
-    $translator.formatPrice(0, subscriptionBasePrice.currency);
-
-  const formattedNonSubscriptionBasePrice =
-    nonSubscriptionBasePrice &&
-    $translator.formatPrice(
-      nonSubscriptionBasePrice.amountMicros,
-      nonSubscriptionBasePrice.currency,
-    );
-
-  let renewalDate = null;
-  const expectedPeriod =
-    subscriptionOption.trial?.period || subscriptionOption.base?.period;
-  if (expectedPeriod) {
-    renewalDate = getNextRenewalDate(
-      new Date(),
-      expectedPeriod,
-      isSubscription,
-    );
-  }
 </script>
 
 <section>
-  <div class="rcb-pricing-info">
+  <div class="rcb-pricing-info" class:has-expanded-details={isSubscription}>
     <ProductInfoHeader
       {productDetails}
       showProductDescription={brandingAppearance?.show_product_description}
     />
 
-    {#if isSubscription}
-      <div class="rcb-product-price-container">
-        {#if subscriptionTrial?.periodDuration}
-          <div class="rcb-product-trial">
-            <Localized
-              key={LocalizationKeys.StatePresentOfferFreeTrialDuration}
-              variables={{
-                trialDuration: getTranslatedPeriodLength(
-                  subscriptionTrial.periodDuration || "",
-                  $translator,
-                ),
-              }}
-            />
-          </div>
-        {/if}
-        <div>
-          {#if subscriptionBasePrice}
-            <span class="rcb-product-price">
-              <Localized
-                key={LocalizationKeys.StatePresentOfferProductPrice}
-                variables={{
-                  productPrice: formattedSubscriptionBasePrice,
-                }}
-              />
-            </span>
-          {/if}
-
-          {#if subscriptionOption?.base?.period}
-            <span class="rcb-product-price-frequency">
-              <span class="rcb-product-price-frequency-text">
-                {$translator.translatePeriodFrequency(
-                  subscriptionOption.base.period.number,
-                  subscriptionOption.base.period.unit,
-                  { useMultipleWords: true },
-                )}</span
-              >
-            </span>
-          {/if}
-        </div>
-      </div>
-
-      <div class="rcb-product-details expanded">
-        <div class="rcb-product-details-padding">
-          {#if subscriptionTrial?.periodDuration}
-            <div class="rcb-product-trial-explanation">
-              <div class="rcb-after-trial-ends rcb-text-dark">
-                <Localized
-                  key={LocalizationKeys.StatePresentOfferPriceAfterFreeTrial}
-                  variables={{
-                    renewalDate:
-                      renewalDate &&
-                      $translator.translateDate(renewalDate, {
-                        dateStyle: "medium",
-                      }),
-                  }}
-                />
-              </div>
-              <div class="rcb-after-trial-ends rcb-text-dark">
-                {formattedSubscriptionBasePrice}
-              </div>
-            </div>
-            <div class="rcb-product-trial-explanation">
-              <div class="rcb-text-dark rcb-total-due-today">
-                <Localized
-                  key={LocalizationKeys.StatePresentOfferPriceTotalDueToday}
-                />
-              </div>
-              <div class="rcb-text-dark rcb-total-due-today">
-                {formattedZeroPrice}
-              </div>
-            </div>
-          {/if}
-        </div>
-      </div>
-    {/if}
-    {#if !isSubscription}
-      <span class="rcb-product-price">
-        <Localized
-          key={LocalizationKeys.StatePresentOfferProductPrice}
-          variables={{ productPrice: formattedNonSubscriptionBasePrice }}
-        />
-      </span>
-
-      {#if brandingAppearance?.show_product_description}
-        <span class="rcb-product-description">
-          <Localized
-            key={LocalizationKeys.StatePresentOfferProductDescription}
-            variables={{ productDescription: productDetails.description }}
-          />
-        </span>
-      {/if}
-    {/if}
+    <ProductInfoPricing
+      {productDetails}
+      {purchaseOption}
+      showProductDescription={brandingAppearance?.show_product_description}
+    />
   </div>
 </section>
 
@@ -176,54 +39,6 @@
     font: var(--rc-text-body1-mobile);
     gap: var(--rc-spacing-gapLarge-mobile);
     user-select: none;
-  }
-
-  .rcb-product-price {
-    color: var(--rc-color-grey-text-dark);
-    font: var(--rc-text-titleMedium-mobile);
-  }
-
-  .rcb-product-trial {
-    color: var(--rc-color-grey-text-dark);
-    font: var(--rc-text-titleLarge-mobile);
-  }
-
-  .rcb-product-price-frequency {
-    color: var(--rc-color-grey-text-dark);
-    font: var(--rc-text-body1-mobile);
-  }
-
-  .rcb-product-price-frequency-text {
-    white-space: nowrap;
-  }
-
-  .rcb-product-details {
-    color: var(--rc-color-grey-text-light);
-    margin: 0;
-    overflow: hidden;
-    transition: max-height 0.2s ease-in-out;
-  }
-
-  .rcb-product-details-padding {
-    display: flex;
-    flex-direction: column;
-    gap: var(--rc-spacing-gapSmall-mobile);
-  }
-
-  .rcb-product-trial-explanation {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-between;
-  }
-
-  .rcb-text-dark {
-    color: var(--rc-color-grey-text-dark);
-    font-weight: 500;
-  }
-
-  .only-desktop {
-    display: none;
   }
 
   .rcb-pricing-info {
@@ -237,50 +52,13 @@
   }
 
   @container layout-query-container (width >= 768px) {
-    .only-mobile {
-      display: none;
-    }
-
-    .rcb-product-details-padding {
-      gap: var(--rc-spacing-gapMedium-desktop);
-    }
-
-    .only-desktop {
-      display: block;
-    }
-
     .rcb-pricing-info {
       margin-top: calc(var(--rc-spacing-gapXXLarge-desktop) * 2);
       gap: var(--rc-spacing-gapXXXLarge-desktop);
     }
 
-    .rcb-pricing-info:has(.rcb-product-details.expanded) {
+    .rcb-pricing-info.has-expanded-details {
       gap: var(--rc-spacing-gapXXXLarge-desktop);
-    }
-
-    .rcb-product-price {
-      font: var(--rc-text-titleMedium-desktop);
-    }
-
-    .rcb-product-price-frequency {
-      font: var(--rc-text-body1-desktop);
-    }
-
-    .rcb-product-details {
-      max-height: 500px;
-      gap: var(--rc-spacing-gapXLarge-desktop);
-    }
-
-    .rcb-total-due-today {
-      font: var(--rc-text-titleMedium-desktop);
-    }
-
-    .rcb-after-trial-ends {
-      font: var(--rc-text-body1-desktop);
-    }
-
-    .rcb-product-trial {
-      font: var(--rc-text-titleLarge-desktop);
     }
   }
 </style>
